Moved cleanString into core and updated cgac_code check to use that
